+++ ZucchiniMVC/Areas/Identity/Pages/Account/Manage/Index.cshtml.cs
@@ -5,8 +5,8 @@ using Microsoft.AspNetCore.Identity.UI.Services;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.AspNetCore.WebUtilities;
-using SharedLib.DTOs.QueuePublisherDOTs;
-using SharedLib.QueuePublisher;
+using SharedLib.DTOs.QueuePublisherDTOs;
+using SharedLib.QueuePublishier;
 using System.ComponentModel.DataAnnotations;
 using System.Text;
 using System.Text.Encodings.Web;
